Refactor providers
