+++ src/login/mod.rs
@@ -1,6 +1,5 @@
 mod tools;
 pub mod providers;
-pub mod provider;
 pub mod services;
 pub mod routes;
 pub mod models;
