+++ src/login/providers/github.rs
@@ -1,5 +1,5 @@
 use async_trait::async_trait;
-use crate::login::provider::{LoginProvider, ThirdPartyUserInfo};
+use crate::login::providers::{LoginProvider, ThirdPartyUserInfo};
 use crate::login::tools;
 use actix_web::Error;
 use crate::env;
